requirements-graphql.txt
fastapi
uvicorn[standard]
strawberry-graphql
psycopg2-binary
